read: make BufProcessor depend on WantsProcessing
We recently introduced `source::WantsProcessing` for communicating
whether an entity, primarily a `Source`, will want to be driven. We
decided, nearly on a whim, that we'd just re-use the trait as the
interface through which a `BufProcessor`, and ultimately an `Item`
signals the same to the `Ring`.

This implies that an `Item` can only be built from a `BufProcessor`
which is only `WantsProcessing`. Since the whole point of `BufProcessor`
is more or less its use in an `Item`, we decided to make `BufProcessor`
always depend on `source::WantsProcessing`.