read: make Ring::prepare only prepare Items that want processing
We currently poll every value we use in every main loop iteration, which
means we usually poll them every second. While this may be just a bit
eager for things like PSI and load, which change every other second, it
really is a waste of CPU time for some other values, most notably a
battery's `charge_full` value. And while making a decision does require
some amount of work in itself, it will most certainly not compare to any
system interaction we need to do, or parsing any value we read.

This change finally introduces the actual filtering of events that will
make it into the IO uring, and thus be processed.