main: move cfg dependent import of Instant to root
We usually import `std::time::Instant`, but we use a mock version for
tests, i.e. we import different `Istant`s depending on the config.
Naturally, this also affects fn parameters, return types and other parts
of the module's public interface, if we use `Instant` in them. This is
not the case yet, but we plan to introduce interface which _will_ use
that type. And due to that, we need to make sure to use the same
`Interface` type in any module using those interfaces.

Because we don't want to duplicate the noisy import, we decided to make
the `Import` availible throughout the crate by moving it to `main`. We
could also just have made it public in `source`, but we don't feel that
having `Instant` imported from some random module just because that
defines some interface that happens to use that type is particularly
reasonable.