source: add Gated Source
Up until now, all `Source`s we have are polled in, more or less, regular
intervals. However, we do have situations where we don't actually have
any use for a new value from a `Source`. There's no point in polling the
current coming from a battery if we already now it's neither charging
nor discharging.

This change introduces a `Source` that only `WantsProcessing` if a given
condition is true, allowing us to reduce the number of reading we need
to do.