spec: reorganize computation of status display in apply_battery
When `Discharging`, the battery status is meant to display an estimate
of the remaining battery lifetime instead of the status symbol. While
this did work previously, we did not handle the absence of an estimate
well. In that case, we would not display the discharging symbol but the
standart fallback.

This wasn't much of an issue for us, since the unavailability of an
estimate was very unlikely, only occuring if the kernel would somehow
give us a value that we could not parse. However, we are planning to
gate updates of the current average based on the status, leading to
estimates not being availible at the instant when the status of a
battery changes to `Discharging`.

This change resolves the issue, making sure we always display something
as long as we have a status. We also took this opportunity to use
standart symbols (albeit lower-case) for single-letter variables in the
closure calculating the estimate.