read: make PSI BufProcessor wrap a Source
When we made `Source`s more composable and introduced the `BufProcessor`
`read::Word` (now `read::Simple`), we decided against having `PSI` wrap
a `Source` but make it hold the data directly. We chose to do so because
we deemed it too much of a hussle when we knew we'd only ever wrap an
`Option<f32>`.

Later, we wanted every `Source` to impl `WantsProcessing`, and so we
also implemented it for `PSI`. Which meant it no longer held just an
`Option<f32>`, essentially voiding the reason we had for its design.
And the `WantsProcessing` impl had a hard-coded interval.

This change now does away with all that, and makes `PSI` wrap an
`Updateable` `Source`, just like `Simple`. Which means we can now just
wrap a `source::LowerRate` rather than duplicating its implementation.